backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
pywebpush==2.0.0
twilio==9.0.5
aiosmtplib==3.0.1
email-validator==2.1.1
pydantic-settings==2.2.1
py-vapid==1.9.0
